feat(website): extend scroll-linked weave
