+++ apps/networkcanvas.com/components/get-started/AppChoiceCard.tsx
@@ -108,6 +108,7 @@ export function AppChoiceCard({ app }: { app: AppRecord }) {
 
   return (
     <article
+      data-get-started-weave-interactive-target
       className={cn(
         'tablet-portrait:p-9 flex h-full flex-col rounded p-7 transition-transform focus-within:-translate-y-1 hover:-translate-y-1 motion-reduce:transform-none',
         treatmentClasses[app.treatment],
